Removed atom vote gossip/sync
Atom votes are 100% ephemeral and are only needed to sync atom mem
pools.

Syncing validators don't need to know the current state of the mem pool.
Providing the network is healthy with 2f+1 vote power progress will
still be made, as they will accept atoms in blocks with super majority
vote as a representation of the atom mem pool being synced.